+++ examples/bouncy_balloon_example.rs
@@ -1,5 +1,6 @@
 use bevy::prelude::*;
 use bevy_inspector_egui::quick::{ResourceInspectorPlugin, WorldInspectorPlugin};
+use bevy_rapier3d::prelude::*;
 use bevy_silk::prelude::*;
 
 mod camera_plugin;
@@ -12,6 +13,7 @@ fn main() {
         })
         .add_plugins(DefaultPlugins)
         .add_plugin(WorldInspectorPlugin)
+        .add_plugin(RapierPhysicsPlugin::<NoUserData>::default())
         .add_plugin(ResourceInspectorPlugin::<ClothConfig>::new())
         .add_plugin(camera_plugin::CameraPlugin)
         .add_plugin(ClothPlugin)
@@ -30,25 +32,18 @@ fn setup(
     mut meshes: ResMut<Assets<Mesh>>,
 ) {
     commands.spawn(DirectionalLightBundle::default());
-    let mesh_handle = meshes.add(shape::Cube::new(1.0).into());
-    [
-        (Color::BLUE, [-10.0, 0.0]),
-        (Color::GREEN, [10.0, 0.0]),
-        (Color::YELLOW, [0.0, -10.0]),
-        (Color::RED, [0.0, 10.0]),
-    ]
-    .map(|(color, [x, z])| {
-        commands.spawn(PbrBundle {
-            mesh: mesh_handle.clone(),
-            transform: Transform::from_xyz(x, 0.0, z),
-            material: materials.add(StandardMaterial {
-                base_color: color,
-                double_sided: true,
-                ..Default::default()
-            }),
+    let mesh = meshes.add(shape::Cube::new(10.0).into());
+
+    // Ground
+    commands.spawn((
+        PbrBundle {
+            mesh,
+            material: materials.add(Color::WHITE.into()),
             ..Default::default()
-        });
-    });
+        },
+        Name::new("Ground"),
+        Collider::cuboid(5.0, 5.0, 5.0),
+    ));
 }
 
 fn spawn_cloth(
@@ -66,10 +61,11 @@ fn spawn_cloth(
                 .into(),
             ),
             material: materials.add(Color::YELLOW.into()),
-            transform: Transform::from_xyz(0.0, 2.0, 0.0),
+            transform: Transform::from_xyz(0.0, 15.0, 0.0),
             ..Default::default()
         },
-        ClothBuilder::new().with_pinned_vertex_ids(0..=0),
+        ClothBuilder::new(),
+        ClothCollider::default(),
         Name::new("Balloon"),
     ));
 }