+++ src/systems/mod.rs
@@ -1,4 +1,3 @@
 /// Base cloth systems
 pub mod cloth;
-#[cfg(feature = "rapier_collisions")]
 pub mod collisions;