+++ src/lib.rs
@@ -1,8 +1,4 @@
-//! # Bevy Silk
-//!
 //! CPU driven Cloth engine for Bevy using Verlet integration.
-//!
-//! by [Félix Lescaudey de Maneville](https://linktree.com/ManevilleF)
 //!
 //! ## Get started
 //!
@@ -10,7 +6,7 @@
 //!
 //! Add `bevy_silk` as a dependency in the `Cargo.toml`
 //!
-//! `bevy_silk = "0.6"`
+//! `bevy_silk = "0.7"`
 //!
 //! Or follow the main git branch
 //!
@@ -26,6 +22,7 @@
 //! | 0.4.0  | 0.9  |
 //! | 0.5.0  | 0.10 |
 //! | 0.6.0  | 0.11 |
+//! | 0.7.0  | 0.12 |
 //!
 //! ### Plugin
 //!
@@ -118,16 +115,20 @@
 //!         ))
 //!         .id();
 //!     let anchor_to_a = VertexAnchor {
-//!         custom_target: Some(entity_a), // The anchor will pin the vertices to `entity_a`
-//!         custom_offset: Some(Vec3::new(1.0, 1.2, 0.0)),  // Specify an extra offset from the target's `GlobalTransform`
+//!         // The anchor will pin the vertices to `entity_a`
+//!         custom_target: Some(entity_a),
+//!         // Specify an extra offset from the target's `GlobalTransform`
+//!         custom_offset: Some(Vec3::new(1.0, 1.2, 0.0)),
 //!         ..Default::default()
 //!     };
 //!     let anchor_to_self = VertexAnchor {
-//!         custom_target: None,  // The anchor will pin the cloth entity
-//!         custom_offset: Some(Vec3::new(-1.0, 0.0, -0.1)), // Specify an extra offset from the target's `GlobalTransform`
+//!         // The anchor will pin the cloth entity
+//!         custom_target: None,
+//!         // Specify an extra offset from the target's `GlobalTransform`
+//!         custom_offset: Some(Vec3::new(-1.0, 0.0, -0.1)),
 //!         ..Default::default()
 //!     };
-//!    
+//!
 //!     let cloth = ClothBuilder::new()
 //!         // Adds pinned vertices ids using an Iterator
 //!         .with_anchored_vertex_ids(0..9, anchor_to_a)
@@ -143,11 +144,12 @@
 //! ```
 //!
 //! Custom anchoring allows to :
-//! - pin vertices to various entities, like skeletal mesh joints
-//! - define custom offsets to customize the distance between the anchored
+//!
+//! * pin vertices to various entities, like skeletal mesh joints
+//! * define custom offsets to customize the distance between the anchored
 //!   vertices an the target
-//! - use world space pinning and ignore the target's rotation for example
-//! - override the vertex positions, using only the offset
+//! * use world space pinning and ignore the target's rotation for example
+//! * override the vertex positions, using only the offset
 //!
 //! ### Configuration
 //!
@@ -179,8 +181,9 @@
 //!
 //! You may add wind forces to the simulation for a more dynamic clothing
 //! effect, for each force you may choose from:
-//! - `Wind::Constant` for constant wind force
-//! - `Wind::SinWave` for a sin wave following wind intensity with custom force
+//!
+//! * `Wind::Constant` for constant wind force
+//! * `Wind::SinWave` for a sin wave following wind intensity with custom force
 //!   and frequency.
 //!
 //! `Wind` forces can be added as a resource to your app through the `Winds`
@@ -240,10 +243,10 @@
 //!
 //! Three `bevy_rapier` components will be automatically inserted:
 //!
-//! - a `RigidBody::KinematicPositionBased`
-//! - a `Collider` which will be updated every frame to follow the cloth bounds
+//! * a `RigidBody::KinematicPositionBased`
+//! * a `Collider` which will be updated every frame to follow the cloth bounds
 //!   (AABB)
-//! - a `SolverGroup` set to 0 (`Group::NONE`) in everything, avoiding default
+//! * a `SolverGroup` set to 0 (`Group::NONE`) in everything, avoiding default
 //!   collision solving.
 //!
 //! You can customize what collisions will be checked by specifying
@@ -271,10 +274,11 @@
 //! ```
 //!
 //! Three `bevy_xpbd` components will be automatically inserted:
-//! - a `RigidBody::Kinematic`
-//! - a `Collider` which will be updated every frame to follow the cloth bounds
+//!
+//! * a `RigidBody::Kinematic`
+//! * a `Collider` which will be updated every frame to follow the cloth bounds
 //!   (AABB)
-//! - a `Sensor` used for avoiding default collision solving.
+//! * a `Sensor` used for avoiding default collision solving.
 //!
 //! You can customize what collisions will be checked by specifying
 //! `CollisionLayers`. (See the [`bevy_xpbd` docs](https://docs.rs/bevy_xpbd_3d/latest/bevy_xpbd_3d/components/struct.CollisionLayers.html)).
@@ -286,19 +290,19 @@
 //!
 //! ## Q&A
 //!
-//! - `My mesh falls immediately and infinitely when I add a Cloth component,
+//! * `My mesh falls immediately and infinitely when I add a Cloth component,
 //!   how to fix it?`
 //!
-//! You probably didn't specify any *pinned points*, meaning there are no
-//! vertices anchored to your entity's `GlobalTransform`.
+//!     You probably didn't specify any *pinned points*, meaning there are no
+//!     vertices anchored to your entity's `GlobalTransform`.
 //!
-//! - `My cloth jitters a lot/ suddenly falls down/ has strange sudden
+//! * `My cloth jitters a lot/ suddenly falls down/ has strange sudden
 //!   behaviour`
 //!
-//! Gravity and winds are by default smoothed out by the framerate, if the
-//! framerate drops suddenly gravity and wind get much stronger.
-//! If your simulation suffers from this you can specify a custom smooth value
-//! in `ClothConfig::acceleration_smoothing`.
+//!     Gravity and winds are by default smoothed out by the framerate, if the
+//!     framerate drops suddenly gravity and wind get much stronger.
+//!     If your simulation suffers from this you can specify a custom smooth
+//!     value in `ClothConfig::acceleration_smoothing`.
 //!
 //! [`bevy_rapier`]: https://github.com/dimforge/bevy_rapier
 //! [`bevy_xpbd`]: https://github.com/Jondolf/bevy_xpbd